Rename workspace-navigation.pure.ts to prepare-workspace-tab.ts
