+++ packages/app/src/utils/prepare-workspace-tab.ts
